v0.3
v0.3

v0.3

v0.3

+++ src/main/java/com/thl/spring/controller/SneakersRestController.java
@@ -4,10 +4,12 @@ package com.thl.spring.controller;
 import com.thl.spring.model.Sneakers;
 import com.thl.spring.service.SneakersService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @RequestMapping("/sneakers")
@@ -18,22 +20,40 @@ public class SneakersRestController {
 
     @PostMapping("/save")
     public ResponseEntity<Sneakers> save(@RequestBody Sneakers sneakers) {
-        return sneakersService.save(sneakers);
+
+        if (sneakersService.isExists(sneakers)) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
+        sneakersService.save(sneakers);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @GetMapping("/findById/{id}")
     public ResponseEntity<Sneakers> findById(@PathVariable("id") int id) {
-        return sneakersService.findById(id);
+        Optional<Sneakers> optionalSneakers = sneakersService.findById(id);
+
+        if (optionalSneakers.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(optionalSneakers.get());
     }
 
     @GetMapping("/findByFirm/{firm}")
     public ResponseEntity<List<Sneakers>> findByFirm(@PathVariable("firm") String firm) {
-        return sneakersService.findByFirm(firm);
+
+        if (!sneakersService.isExistsByFirm(firm)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(sneakersService.findByFirm(firm));
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<?> delete(@PathVariable("id") int id) {
-        return sneakersService.delete(id);
+    public ResponseEntity<Integer> delete(@PathVariable int id) {
+        if (!sneakersService.isExistsById(id)) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        sneakersService.delete(id);
+        return ResponseEntity.status(HttpStatus.OK).body(id);
     }
 
 }
